Add old input system + add prettier WebGL template
Game now uses both new and old input systems to move character. Could potentially cause problems but seems to work. The only new code that was added was the Update() in Runner.cs script, the rest of the game was untouched.
Also added a WebGL template to remove ugly Unity bar on bottom of web-playable version.